Fix graph module on OSX
